Repair main branch Swift Testing expressions blocking focused validation
